+++ custom_components/tibber_data/application_credentials.py
@@ -23,6 +23,8 @@ async def async_get_authorization_server(hass: HomeAssistant) -> AuthorizationSe
     )
 
 
+
+
 async def async_get_description_placeholders(hass: HomeAssistant) -> Dict[str, str]:
     """Return description placeholders for OAuth2 setup."""
     return {